Move extractor factory into its own file, adjust test accordingly
